shader_runner: Support optional internal format for 'texture rgbw'
The motivation for this is that (as noted by curro)
ARB_shader_image_load_store doesn't support images with an internal
format of GL_RGBA.

See 'Table X.2' in
https://www.opengl.org/registry/specs/ARB/shader_image_load_store.txt

Therefore, in order to use image load/store in shader runner tests, we
must override the default internal format when creating the image.

We retain the default internal format of GL_RGBA when the optional
parameter is not given.

Now an internal format can optionally be added as a new parameter when
creating the image. For example:

  texture rgbw 0 (16, 16) GL_RGBA8

v2:
 * Drop 'base type' parameter

Signed-off-by: Jordan Justen <jordan.l.justen@intel.com>
Reviewed-by: Francisco Jerez <currojerez@riseup.net>