aa: too sensitive at the line end points for PNV platform
Adjust middles of the segments sample so that now the test pass also
on PNV platform.

Bspec documentation G45: Volume 2: 3D/Media
Last Pixel Enable  bit of  SF_STATE controls whether the last pixel
of the last line in a LINESTRIP_xxx primitive
or the last pixel of each line in a LINELIST_xxx primitive is rendered.

Signed-off-by: Marius Predut <marius.predut@intel.com>
Reviewed-By: Eduardo Lima Mitev <elima@igalia.com>