framework/summary.py: raise a FatalError for name collision
When two json results have the same "name" value in the root of the json
dictionary, piglit will try to create two directories with the same
name. In this event an OSError will be raised to the top level exception
handler, resulting in a bug warning.

This patch raises an expected error, giving the user advice on resolving
the issue.

Signed-off-by: Dylan Baker <dylanx.c.baker@intel.com>
Tested-by: Ben Widawsky <ben@bwidawsk.net>