sampler_array_indexing: also check that texture coordinates work
Previously the shaders were using a checkerboard texture with
identical colours for each part of the grid.
That meant that it didn't actually matter what texture coordinates
were used to sample. This patch changes it so that they all have a
dark red colour for one of the colours and then it ensures that the
texture coordinates should sample from a space with the other colour.

Test results:
Nvidia GeForce 840M - NVIDIA 346.47: pass

Reviewed-by: Neil Roberts <neil@linux.intel.com>