arb_gpu_shader5: Fix rounding instability in UBO and sampler indexing vs-nonuniform-control-flow tests.
Fixes a rounding instability that would cause shader_runner to probe
pixels offset by one for some points close to the right edge of the
window on systems using x87 floating point arithmetic with certain
compiler versions (the test seemed to work fine when built with GCC
v5.1.0, but failed on some systems with GCC v4.9.2).

The reason for the instability was that the default window height and
width of 250 pixels was evenly divisible by all fractions used as
point coordinates, what would cause the coordinates passed to relative
rgba probes to lie precisely on the boundary between four pixels,
giving unpredictable results in presence of the slightest rounding
error.

Instead use a coprime of 10 as window size to guarantee that there's
always one fragment closer than the other three.  The 1/250 half-pixel
offset previously used in the vertex shader now becomes unnecessary
because the different fragment locations are sufficient to guarantee
consistent rasterization results.

This gets rid of the points which had one of the coordinates equal to
0 and replaces them with points close to the top and right edges
because they would have necessarily been at a half-integer distance
from the closest fragments regardless of the window size.

Bugzilla: https://bugs.freedesktop.org/show_bug.cgi?id=90993
Reported-by: Mark Janes <mark.a.janes@intel.com>
Tested-by: Mark Janes <mark.a.janes@intel.com>