framework/tests/utils.py: fix set_env decorator
This decorator sets environment variables, then runs the test, then
restores them.

If the value is stored as __DONOTRESTORE__, the idea is that it should
be deleted, otherwise it should be restored. But, before this patch it
would set the value to __DONOTRESTORE__ if the value was
__DONOTRESTORE__, and the key wasn't in os.environ.

This fixes that.

Signed-off-by: Dylan Baker <dylanx.c.baker@intel.com>