framework: Re-raise uncaught exceptions from their original context
When an uncaught exception reaches the top level exception handler, and
PIGLIT_DEBUG is true, then the exception will be re-raised. The problem
is that currently the exception is re-raised as 'raise e', which raise
the correct exception, but does so from the wrong context, making the
back-trace completely useless. By re-raising by simply using the 'raise'
keyword causes the exception to be passed on 'as-is', making the back
trace useful for debugging.

Trivial.

Signed-off-by: Dylan Baker <dylanx.c.baker@intel.com>