xts: Add a set of rendercheck tests to run.
We don't do a full run, because that takes approximately an eternity.
However, running a small subset can be useful to catch certain bugs in
Render implementations.  This passes on an fb implementation in <4
seconds on my laptop.

v2: Style fixes from Dylan (one variant on his suggestions, to avoid
    grouptools.from_path).

Reviewed-by: Dylan Baker <baker.dylan.c@gmail.com> (v1)
Reviewed-by: Dave Airlie <airlied@redhat.com> (v1)