framework/programs/run.py: On a new run clear results directory
A current source of confusion right now, is to run piglit, and setting a
results directory that already contains a partial result in it. If the
new run is shorter than the run it's overwriting, then when it
accumulates all of the tests together into a single file at the end of
the run, these extra tests from the previous run will be included.

To prevent that, this patch clears the contents of the directory (at an
implementation level it actually deletes and recreates the directory),
so that there is a clean working directory to work with.

Obviously this doesn't affect the resume path.

Signed-off-by: Dylan Baker <dylanx.c.baker@intel.com>