compression.py: fix compression to be updated when piglit.conf is reloaded
There is currently a bug in piglit (demonstrated by the previous patch),
that shows that compression is hard set before piglit.conf is loaded in
run, and not affected by any subsequent re-reads of piglit.conf. This is
problematic, resulting in the compression mode being either
PIGLIT_COMPRESSION or the default value, bz2.

This patch corrects that by removing constant values and replacing them
with a getting function.

Signed-off-by: Dylan Baker <dylanx.c.baker@intel.com>