compressed_backend_tests.py: add test to demonstrate bug in compression
Currently the compression code has no way to handle a change in
piglit.conf (which incidentally it needs to, since that happens
immediately in run mode). This results in the default mode (bz2) being
picked when a PIGLIT_COMPRESSION environment variable isn't specified,
even when a piglit.conf value is.

Signed-off-by: Dylan Baker <dylanx.c.baker@intel.com>