framework tests: fix binary_check utility bug
This patch changes binary_check to work when a binary is present, but
doesn't work for some reason.

The example here is in the json content tests, where it looks for
glxinfo, but if X isn't running glxinfo provides no output, which will
cause a test to fail. This isn't useful, so instead of using the *nix
command 'which', call the command and allow an expected returncode to be
passed, if the returncode isn't that returncode, then skip.

Signed-off-by: Dylan Baker <dylanx.c.baker@intel.com>