framework: when searching a directory ignore files ending in .old
Currently if piglit updates a result to a new version, then tries to
load from the same directory the first result will be
'results.json.old', and piglit will interpret the '.old' as a
compression extension, and die in a fire.

With this patch when searching a folder (but not when pointing directly
at a file) piglit will ignore any files ending in '.old', which will
correct the problem.

A unit test is also included.

Signed-off-by: Dylan Baker <dylanx.c.baker@intel.com>