arb_shader_atomic_counters-semantics: test different binding points
Piglit was only checking the expected effects using binding point 0.
With this commit, all the values between 0 and
GL_MAX_ATOMIC_COUNTER_BUFFER_BINDINGS are checked.

It also checks that the binding point used on the shader and the one
returned by
glGetActiveAtomicCounterBufferiv(..., ...,
GL_ATOMIC_COUNTER_BUFFER_BINDING, ...) is the same.

This test detects mesa bug:
https://bugs.freedesktop.org/show_bug.cgi?id=90175