xz: Use --help when detecting xz presence
If xz is present, and the piglit command's output is redirected, then
xz will think its output is redirected. This will cause xz to try to
read data from stdin to compress.

Instead we can run xz with --help to cause it to print help
information if the xz executable is present, and prevent xz from
trying to compress data from stdin.

Since we don't want to see 'xz --help' output from piglit, we now need
to redirect both stdout and stderr to /dev/null.

This allows us to pipe the output of piglit summary when using the shell
xz compressor.

v2 (Dylan):
- send both stdout and stderr directly to /dev/null
- Fix xz detection in the xz shell path.

Signed-off-by: Jordan Justen <jordan.l.justen@intel.com>
Signed-off-by: Dylan Baker <baker.dylan.c@gmail.com>