framework: move unicode conversion handling to TestResult
This implements a python descriptor pattern to handle setting the out
and err attributes, which should always be unicode.

The familiar "property" decorator is an example of a descriptor,
although it is a more terse way of doing it. And while that could have
been done here, by not making it a property we can reuse the exact same
class for both out and err, and could theoretically use it for other
attributes.

Signed-off-by: Dylan Baker <dylanx.c.baker@intel.com>