arb_shader_image_size/builtin: test all the image formats
This patch introduces the concept of stage and image format of interest in
order to reduce the number of subtests actually being tested.

The quick mode's behaviour is unchanged while we also introduce the slow mode
that forces running the tests over all the image formats and stages.

The slow mode takes 6.5 minutes on a Fermi with the NVIDIA proprietary driver
while the normal execution time is around 1 second.

Signed-off-by: Martin Peres <martin.peres@linux.intel.com>
Reviewed-by: Francisco Jerez <currojerez@riseup.net>