framework/results: make the result of a test the worst of its subtests
This patch changes the behavior of the results object, so that when
results.object is called the worst subtest status is returned as the
result, if there are subtests. If there are not subtests the behavior is
not changed.

This is currently done manually in the summary code, but with this
change that will no longer be required.

Signed-off-by: Dylan Baker <dylanx.c.baker@intel.com>