framework/results.py: Add a method to TestrunResult to calculate totals
Currently this is done in the summary code, and is very complicated.
This change has a couple of advantages. First, having it in the
TestrunResult object seems a more obvious place for it. Second, this
will allows us to store the data in the json, removing the need to
calculate the data again and again.

Signed-off-by: Dylan Baker <dylanx.c.baker@intel.com>