framework/results.py: Add a to_json method to the TestrunResults
The piglit_encoder knows how to convert an object with a to_json method
into it's json representation. By using this approach we can add special
hints to the json for converting it back into it's python object
representation, which can also be removed easily by using a from*
method.

Signed-off-by: Dylan Baker <dylanx.c.baker@intel.com>