framework/summary.py: Make the text mode less weird
Describing the current implementation as anything but weird is hard. The
list of tests and statuses it generates is pretty reasonable, but the
summary is very weird. It counts more than one result in some cases, but
not in others (for comparison statuses), and generates a single column
of data. This is confusing and difficult to read at best.

This rework instead creates a column (each one will be 20 characters)
for each result and properly calculates both the number of tests in each
category, and the number of changes (and all subsets of changes).

Signed-off-by: Dylan Baker <dylanx.c.baker@intel.com>