framework/summary.py: Add a class to summary for calculating categories
This class lazily evaluates regressions, problems, etc; and caches the
results. This generates a set or list of sets containing the names of
the tests, and a count of the number of tests.

This also adds a new decorator, the lazy_property. This is a very clever
Descriptor class, that adds the calculated value to the instance (while
the descriptor itself belongs to the class) under the same name, thus
shadowing itself. This means that the calculation of the Descriptor is
done once, and then stored, making any method decorated with this class
lazy.

Signed-off-by: Dylan Baker <dylanx.c.baker@intel.com>