framework/summary.py: split console generator from Summary object
The summary object is a giant, bloated, awful thing. Using all of the
groundwork we've put together in the previous patches we can now
starting ripping it to pieces, pulling out the parts worth saving, and
converting them to use the new data structures, and throwing away a lot
of awful code in the process.

As a side effect, the list of tests will now include subtests.

Signed-off-by: Dylan Baker <dylanx.c.baker@intel.com>