framework/summary.py: remove the Summary object.
This removes the last vestiges of the Summary object, which wasn't doing
too much other than holding a few constants and turning result files
into a Results object. With this html is a stand alone function making
use of the Results class and it's members, rather than the Summary
class.

Signed-off-by: Dylan Baker <dylanx.c.baker@intel.com>