framework/grouptools: fix join to handle empty elements correctly
Currently join doesn't handle empty values correctly, as it will add an
extra {SEPARATOR} for each of them.

Passing in this value: ['', 'foo', 'bar', '']
will result in this: '@foo@bar@'

This obviously isn't what we want, we really want 'foo@bar', and this
patch corrects the behavior to do that.

Signed-off-by: Dylan Baker <dylanx.c.baker@intel.com>