framework/summary.py: update copyright header
This patch makes two changes:

First, it replaces the license header with a differently formatted, and
very slightly differently worded, version. This version is taken from:
https://opensource.org/licenses/MIT
and is used in most of the piglit framework.

Second, after some archeology I've added copyright assignment to
contributers back to 2013 (at which point the summary module was
completely rewritten).

Signed-off-by: Dylan Baker <dylanx.c.baker@intel.com>