arb_shader_storage_buffer_object: fix TOP_LEVEL_ARRAY_STRIDE expected value for std430 case
struct B has a size of 52. Interface packing layout qualifier std430 says that
the array of structures are not rounded up a multiple of the base alignment of
a vec4 but to the next multiple of the base alignment of the structure.

The base alignment of a structure is N, where N is the largest base
alignment value of any of its members.

In this case the largest base alignment is 8 (which is mat2's base alignment
following std430 rules), so 52 is rounded up to 56, not to 64.

Tested on NVIDIA proprietary driver version 352.21.

Signed-off-by: Samuel Iglesias Gonsalvez <siglesias@igalia.com>