sso: Add a test that passes data using the legacy gl_TexCoord varyings.
In compatiblity profiles, the GL_ARB_separate_shader_objects extension
allows passing data via built-in varyings such as gl_TexCoord[].  We
don't do compatibility profiles, but we do expose SSO in legacy GL
contexts and allow it with GLSL 1.30 and earlier.

This test actually tries to do that in a rendering test.

This is particularly interesting because Mesa's VARYING_SLOT_* enums
handle built-in varyings different than generic ones.  I wanted to be
able to see how those came through; this provides a simple example.

Signed-off-by: Kenneth Graunke <kenneth@whitecape.org>
Reviewed-by: Ian Romanick <ian.d.romanick@intel.com>