framework/log: add an HTTP logger
Add an HTTP logger so that it is possible to monitor the progress and
status of a test run remotely through a web server. The web server runs
on port 8080 by default and can be configured to use a different port by
setting the "port" key in the "http" section of the configuration file.
The server responds to requests for /summary, which provides a summary
of the current state in JSON format.

v2: Add appropriate locking when reading state data for the web server
    Stop the server after the request for the final results
    Remove the full results page
    Add a configuration option to set the server port (Dylan Baker)
v3: Use the new fallback parameter of safe_get (Dylan Baker)
    Convert the test result status back to string
v4: Add newline after BaseHTTPServer imports (Dylan Baker)

Signed-off-by: Thomas Wood <thomas.wood@intel.com>
Reviewed-by: Dylan Baker <baker.dylan.c@gmail.com>