tests: Add integration for khronos CTS runner.
This adds support for running Khronos' deqp-based conformance suite with
piglit. This gives access to all of the powerful features of pigilt,
per-process tests, the junit and json backends, the summary tools, and
the familiar interface, with minimal fuss.

This is a very small change, since it is deqp-based, and piglit has a
framework for handling deqp based suites already.

Reviewed-by: Mark Janes <mark.a.janes@intel.com>
Signed-off-by: Dylan Baker <dylanx.c.baker@intel.com>