arb_shader_atomic_counters/max-counters: Run the combined atomic buffer test in more cases.
The subtest with at most the maximum number of atomic counter buffers
can be run whenever the combined limit is at least one more than the
limit for the FS stage.  The subtest exceeding the maximum number of
combined atomic counter buffers can be run whenever the sum of the
limits for the VS and FS stages is greater than the combined limit.

Reviewed-by: Timothy Arceri <t_arceri@yahoo.com.au>