framework: add tox testing support
tox is a python test manager that runs the tests in isolated
environments with their own copies of dependent libraries, and can be
used to test against multiple versions of python. This allows us to
easily test the framework with and without the acceleration modules it
can optionally use (simplejson and lxml), it will also allow us to test
against multiple versions of python 3.x (which we support for the
generators, and presumably will support in the future)

This uses the coverage module, and uses nose's cover-tests option.
Unfortunately nose assumes that anything /[Tt]est/ is a test, and hides
the tests for the test package. This might be resolvable by modifying
the regex, but it isn't something straightforward to fix. Using
--cover-tests will get us the test package, but it also gets us the
tests package.

Signed-off-by: Dylan Baker <dylanx.c.baker@intel.com>