tox: Also test python 3.{3-5} for the generators.
This tests the test generators with python 3.x as well as 2.x. The
generators have been supposed to build with python 3.3+ for some time,
but haven't.

This exposes a bug in the tessealation generators that keep them from
building with python 3.x, and these tests fail. This is expected, and
there are patches out for review to correct this.

Signed-off-by: Dylan Baker <dylanx.c.baker@intel.com>