arb_gpu_shader_fp64: expand getuniform double coverage
After looking at CTS I realised we have to accept all types
for get uniform for double conversion, and we should do double
conversion on the non-double types.

I assume this is due to it being the only interface to check
if the location has a value or something.

This adds getting non-double values with GetUniformdv, and
getting double values with float/int/uint interfaces.

notes: currently mesa fails this.

Signed-off-by: Dave Airlie <airlied@redhat.com>