framework: add exception back to TestResult class
Like command, this somehow didn't get moved to the new TestResult class.
Generally this isn't a problem since this is a very uncommon path, but
in some cases this does get hit, and it would cause an exception in the
runner threads. Python has very bad exception handling in threads,
namely it doesn't have any; this results in an exception being raised
and remaining uncaught, terminating the thread, which in turn results in
the test result being incomplete, even though it should have been crash.

This patch fixes that situation.

Signed-off-by: Dylan Baker <dylanx.c.baker@intel.com>
Tested-by: Mark Janes <mark.a.janes@intel.com>
Reviewed-by: Mark Janes <mark.a.janes@intel.com>