clipflat: Fix subtest reporting
There are two problems with the way this test reports subtests:

First, it only reports subtests when they fail. This has all kinds of
problems. It's not at all what the summary code expects, the
summary code expects that subtests are always present, which means that
they will show up under disabled/enabled pages instead of
fixes/regressions, which in turn, will confuse developers.

Second, it tries to report lots of information as the subtest name. This
causes the JSON parser in python (JSON is used by the tests to
communicate with the runner) to choke and die.

This patch makes a couple of changes to correct this.
First, it reports the subtest result always, whether pass or fail.
Second, it simplifies the names of the subtests, and passes the other
information to stdout, which the runner will scoop up and record for a
developer to go look at.

v2: - Fix some formatting that wasn't right (indents with 4 spaces not 8
      space tabs)
    - Report all subtests in both pass and fail cases
    - Fix some minor things as reported by Brian Paul
    - simplify some control flow
    - add function to report each quadrant as a separate subtest with
      nice names

Reviewed-by: Brian Paul <brianp@vmware.com>
Signed-off-by: Dylan Baker <dylanx.c.baker@intel.com>