cl: Update integer/long limit tests to detect incorrect storage sizes
The tests for the integer/long minimums do not properly
check that the value is stored in the correct type.
E.g. (-2147483648) actually gets parsed as a long by the preprocessor,
and therefore INT_MIN is actually stored as long and can't be automatically
vectorized.

By subtracting a vector with value of 0 from the given defined *_MIN
and then grabbing the first element of the resulting vector, we can
make sure that the values are actually stored in the correct type.

v2: Remove char/short modifications.
    Test *_MAX vectorization as well as _MIN for int/long

Reviewed-by: Jan Vesely <jan.vesely@rutgers.edu>
Reported-By: Moritz Pflanzer <moritz.pflanzer14@imperial.ac.uk>
Signed-off-by: Aaron Watry <awatry@gmail.com>