glsl-es: Additional constant initializer tests.
Most built-in functions with constant-expression parameters are also
constant expressions.  GLSL ES 1.00 and 3.00 differ as to whether or not
the sequence operator is a constant expression.

Note: spec/glsl-es-1.00/compiler/const-initializer/from-function.* and
spec/glsl-es-1.00/compiler/const-initializer/from-sequence-in-function.*
currently fail on Mesa.  This is very confusing because the nearly
identical spec/glsl-es-3.00/compiler/const-initializer/from-function.*
tests both pass!

Signed-off-by: Ian Romanick <ian.d.romanick@intel.com>