framework: Add a TimeAttribute class
This class will provide a new interface for time, one that stores the
start and stop times, rather than the total time. It then provides
methods for getting the total and the delta (as used in the summary).

This commit adds the new class and tests for the class, but doesn't
start plugging it in yet.

The goal of this change is to allow post-processing to determine which
tests were running in parallel.

Reviewed-by: Mark Janes <mark.a.janes@intel.com>
Signed-off-by: Dylan Baker <dylanx.c.baker@intel.com>