framework: use TimeAttribute for TestResult and TestrunResult
This makes use of the time TimeAttribute for handling times in results,
both for tests and for runs.

This change requires bumping the json version, and a large chunk of the
changes in this patch are for that change.

This could be split into two patches, but that would require making two
sequential bumps to the json results version, and I felt that the larger
patch was better than making two version bumps.

Reviewed-by: Mark Janes <mark.a.janes@intel.com>
Signed-off-by: Dylan Baker <dylanx.c.baker@intel.com>