ext_texture_format_bgra8888: Add test for GL_BGRA format API errors in Tex(Sub)Image
This is a new test that checks valid and invalid combinations of GL_BGRA_EXT
format and internal format in Tex(Sub)Image2D calls, as specified by the
EXT_texture_format_BGRA8888 extension.

Although the test is very basic and doesn't really render anything or check
pixels, it would have prevented the regression described in the bug below,
which was due to an API error. In the future though, more tests should be
added to validate the extension more thoroughly.

Bugzilla: https://bugs.freedesktop.org/show_bug.cgi?id=92265