framework/sumamry/common.py: Fix subtest handling in non-all groups
The gist of the problem is that we handle lookup for subtests in the all
group correctly, but when we went to compare between multiple results we
don't.

This patch makes use of the get_result from the previous patch to
simplify test name lookups, and fix the non-all case. This also fixes
the unit tests introduced in the previous patch.

Signed-off-by: Dylan Baker <dylanx.c.baker@intel.com>
Tested-by: Marek Olšák <marek.olsak@amd.com>