framework/tests/oglconform_tests.py: Add tests for tests/oglconform.py
This adds a pretty extensive set of tests for oglconform.py, with the
goal of making further cleanups and refactors easier and not introducing
regressions.

This adds a dependency of mock for the unittests. No production code
needs this dependency, so this wont affect most (any?) piglit users.
Mock allows for better testing, in fact, it allows oglconform to be
tested even when it's not installed.

Signed-off-by: Dylan Baker <dylanx.c.baker@intel.com>