tests/oglconform.py: minor refactor to simplify try/except block
This moves code out of the try/except block that we don't mean to catch
exceptions in, which is better coding practice.

This also sets the except block to ValueError only, which is the error
that will occur if there aren't enough values to unpack, which the
except block was added to catch.

Signed-off-by: Dylan Baker <dylanx.c.baker@intel.com>