arb_gpu_shader5: add output varying to stream compile test
When arb_enhanced_layouts is added to mesa this validation will be
moved to the checks on each variable it is applied to rather than on the global
itself so we need to add an output.

V2: don't change emit function stream as this is tested in another piglit, add
additional spec quote.