multisample-accuracy: Add an error margin for lit/unlit pixels
The unlit and totally_lit stats are supposed to count the pixels where
either a primitive completely covers the pixel or no primitive touches
it at all. However this is effectively only determined by checking
whether any primitive has intersected one of the supersample positions
of the reference image. It's completely possible for a primitive to
intersect the pixel boundary but completely miss any of the
supersample positions. The GL implementation is free to pick whatever
multisample positions it wants within the pixel boundary so it's also
possible for a primitive to intersect a multisample position but miss
all of the supersample positions of the reference image. To cope with
this this patch now allows a small margin of error for the pixels that
are either fully lit or fully unlit.

This was causing the test to give false negatives for i965 with the
16x MSAA positions because in that case it chooses positions that are
exactly on the pixel boundary and these would be outside of the grid
of positions chosen by the supersampling in the reference image. The
positions used are actually those described by the D3D API so it is
possible that other hardware would have the same problem.