tests/oglconform.py: make interpret_result() faster in most cases
This optimizes the common paths of OGLCTest.interpret_result(): pass and
the common skip case.

For most OGLCTest cases the result will be in the final 3 lines of
stdout. In this case it is substantially faster to split the last 3
lines, and then check for a string in that list than to use either an
RegexObject (even a pre-compiled one) or to use a 'str in str'
approach.

I've used a script to benchmark the performance of each approach, using
the actual output of a test. This assumes a single expression in the
RegexObject (no use of 'or'). This script is in the mailing list archive
for those curious, but I felt it was too long to leave in a commit
message.

The final approach is 10x faster than the 'str in str' approach, and
6.5x faster than the compiled re. While interpret_result doesn't
constitute a huge amount of time in the piglit run, it does account for
some of it, and this should at least shave a few seconds off the final
run time.

Signed-off-by: Dylan Baker <dylanx.c.baker@intel.com>