fbo-depth-array: Don't create an enormous color render target
When the -fbo option is given to the fbo-depth-array test it tries to
render to a range of different depth/stencil buffer sizes with
dimensions of up to 8192. It verifies the rendering was correct by
binding the depth or stencil buffer as a texture source, rendering the
texture into the Piglit winsys fbo and then probing the result. In
order to cope with all of the possible buffer sizes it was setting the
winsys framebuffer size to 8192x8192. Doing glReadPixels on a buffer
this size causes problems for some platforms and makes the test run
extremely slowly. The point of the test is just to check rendering to
depth and stencil array textures of different sizes so the size of the
renderbuffer isn't really relevant. This patch makes it instead leave
the winsys buffer at the default size when an FBO is used and then
render the texture multiple times in slices until it has checked the
entire texture.

Acked-by: Ben Widawsky <benjamin.widawsky@intel.com>