framework/test/opengl.py: Add FastSkipMixin which checks extensions
This Mixin provides a way for OpenGL tests to skip very fast. Currently
it only applies to GL extensions, but will be extended to cover GLSL
version requirements and GL version requirements (and ES)>

This is split into a separate module because it's going to grow into a
fairly large amount of code (mostly around querying wflinfo).

Signed-off-by: Dylan Baker <dylanx.c.baker@intel.com>