framework/test/opengl.py: add support for GL(ES) version skipping.
This patch adds support to the FasSkipMixin for skipping on GL and GLES
versions. It does this by querying wflinfo for the highest supported

This skipping only supports checking that the requirement is <= to the
maximum GL(ES) provided. There are some instances of tests that have
requirements like "GL <= 2.1", and this patch doesn't cover case. These
cases are pretty rare, so it doesn't seem like a path worth optimizing
for anyway.

Signed-off-by: Dylan Baker <dylanx.c.baker@intel.com>