framework: Add support for feature readiness.
This adds a new "summary feature" command to piglit which creates a HTML table
with the feature x DUT status (useful for multiple features, multiple DUTs).
Another use case is the feature status for subsequent test results (depending
on the meaning of that test result - DUT or build)
A feature readiness is defined by the piglit regexp which selects the tests
relevant for that feature and the acceptance percentage threshold (pass rate).

It requires an input json file containing the list of features, in the
following format (this is just an example):
{
    "glsl" : {
        "include_tests" : "glsl",
        "exclude_tests" : "",
        "target_rate" : 90
    },
    "arb" : {
        "include_tests" :  "arb_gpu",
        "exclude_tests" : "",
        "target_rate" : 10
    }

}

v3:
Changed json rate to int instead of string
Applied other review comments

v2:
Apply review comments (Dylan, Thomas)
Fixed 2nd round of review comments

Signed-off-by: Gabriel Feceoru <gabriel.feceoru@intel.com>
Reviewed-by: Dylan Baker <baker.dylan.c@gmail.com>